Reference count AUTHs
RPCSEC GSSv3 has the concept of a parent and a compound credential.  As
the normal course of operation involves using multiple AUTHs per client
connection, and providing parent and compounds AUTHs when creating a
GSSv3 AUTH, we need a way of reference counting them so that
AUTH_DESTROY does not free them out from under a GSSv3 AUTH that is
using them.

Signed-off-by: Steve Dickson <steved@redhat.com>