portmapper: allow TCP-only portmapper
Code that works in GLIBC's runrpc implementation fails with libtirpc.
libtirpc forces the RPC library to talk to the portmapper via UDP,
even when the client specifies TCP.  This breaks existing code which
expect the protocol specified to be honored, even when talking to
portmapper.

This is upstreaming of an old patch by Rob Riggs reported in Fedora.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1725329
Signed-off-by: Rob Riggs <rob+redhat@pangalactic.org>
Signed-off-by: Dan Hork <dan@danny.cz>
Signed-off-by: Steve Dickson <steved@redhat.com>